Fix: not removing ruby on native installation

+++ lib/escualo/ruby.rb
@@ -1,6 +1,11 @@
 module Escualo
   module Ruby
     def self.install(session, options)
+      if options.with_native_ruby
+        puts "[Escualo] Using native ruby. Not installing it."
+        return
+      end
+
       session.tell! 'apt-get purge libruby* -y'
       if options.with_rbenv
         session.tell_all! 'curl https://raw.githubusercontent.com/fesplugas/rbenv-installer/master/bin/rbenv-installer | bash',
@@ -9,8 +14,6 @@ module Escualo
         session.tell_all! 'rbenv install 2.3.1',
                           'rbenv global 2.3.1',
                           'rbenv rehash'
-      elsif options.with_native_ruby
-        puts "[Escualo] Using native ruby. Not installing it."
       else
         Escualo::AptGet.install session, 'ruby2.3 ruby2.3-dev'
       end
